Split PostgreSQL session reads and writes
