+++ packages/agent-store/agent-store.cabal
@@ -51,7 +51,10 @@ library
                     , Agent.Store.Postgres.Custom.Sql
                     , Agent.Store.Postgres.Custom.Types
                     , Agent.Store.Postgres.Skill.Types
+                    , Agent.Store.Postgres.Session.Read
+                    , Agent.Store.Postgres.Session.Schema
                     , Agent.Store.Postgres.Session.Types
+                    , Agent.Store.Postgres.Session.Write
                     , Agent.Store.Postgres.Hasql
                     , Agent.Store.Postgres.Sql
                     , Agent.Store.Postgres.SessionItem
